Add create computer request
