+++ src/components/Computer/Add-computer/useform.js
@@ -3,21 +3,11 @@ import { useState, useEffect } from 'react'
 
 function useForm(callback, validate, currentComputer) {
 
-  const [computer, setComputer] = useState( { id: null, name: '', introduced: '', discontinued: '', companyDTO: { id: null, name: '' } });
-  const [companyDTO, setCompanyDTO] = useState({ id: null, name: '' });
+  const [computer, setComputer] = useState( { id: null, name: '', introduced: null, discontinued: null, companyDTO: { id: null, name: '' } });
   const [errors, setErrors] = useState({});
   const [isSubmitting, setIsSubmitting] = useState(false);
 
-  function getCompanyDTO(event) {
-    const companyName = event.target.value;
-    const idCompany = event.target.options.selectedIndex;
-    const companyDTOX = { id: idCompany, name: companyName };
-    setComputer({ ...computer, companyDTO: companyDTOX })
-  }
-
   function handleSubmit(event) {
-    console.log(computer)
-    let formIsValid = false;
     event.preventDefault()
     setErrors(validate(computer));
     setIsSubmitting(true);
@@ -38,7 +28,6 @@ function useForm(callback, validate, currentComputer) {
 
   return {
     handleSubmit,
-    getCompanyDTO,
     computer,
     errors
   };
