+++ src/containers/computer/Computers.hook.js
@@ -1,4 +1,4 @@
-import {getComputers, deleteComputer} from '../../api/computer.api'
+import {getComputers, deleteComputer, createComputer} from '../../api/computer.api'
 
 export async function getComputer(page){
     return await getComputers(page)
@@ -6,6 +6,10 @@ export async function getComputer(page){
 
 export async function deleteComputers(tableau){
     return await deleteComputer(tableau)
+}
+
+export async function addComputer(computer){
+    return await createComputer(computer)
 }
